+++ target.json
@@ -1,0 +1,5 @@
+{
+  "baseUrl": "http://localhost/addressbook/",
+  "username": "admin",
+  "password": "secret"
+}
